build: make the project fit the android project structure and buildable
