--- AndroidManifest.xml
@@ -1,14 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.woopra.android.sdk"
-    android:versionCode="1"
-    android:versionName="1.0" >
-
-    <uses-sdk
-        android:minSdkVersion="9"
-        android:targetSdkVersion="19" />
-
-    <application
-        android:label="@string/app_name" >
-    </application>
-
-</manifest>
