--- src/com/woopra/tracking/android/Constants.java
@@ -1,15 +1,0 @@
-package com.woopra.tracking.android;
-
-/**
- * Created by dev on 4/12/17.
- */
-
-public class Constants {
-
-    public static final String W_TRACK_ENDPOINT = "http://www.woopra.com/track/ce/";
-    public static final String W_IDENTIFY_ENDPOINT = "http://www.woopra.com/track/identify/";
-    public static final String W_PING_ENDPOINT = "http://www.woopra.com/track/ping/";
-    public static final String APP_KEY = "Woopra_android";
-    public static final String COOKIE_KEY = "Woopra_android_cookie";
-    public static final String NOT_SET = "NOT_SET";
-}
